+++ test/utest-offline/testlinkapi_offline_test.py
@@ -188,7 +188,7 @@ SCENARIO_KEYWORDS = {'getTestCasesForTestSuite' : {
                                   {'step_number': '7', 'actions': 'action 7 createTestCaseSteps.create', 'execution_type': '2', 'active': '1', 'id': '8157', 'expected_results': 'create - cause step 7 not yet exist'}, 
                                   {'step_number': '8', 'actions': 'action 8 createTestCaseSteps.update', 'execution_type': '2', 'active': '1', 'id': '8158', 'expected_results': 'create - cause step 8 not yet exist'}], 
                         'author_id': '1'}, 
-                    {'node_order': '1', 'is_open': '1', 
+                                     {'node_order': '1', 'is_open': '1', 
                         'keywords': {'2': {'keyword_id': '2', 'notes': 'another key word', 'testcase_id': '8159', 'keyword': 'KeyWord02'}}, 
                         'id': '8159', 'node_type_id': '3', 'layout': '1', 'tc_external_id': '3', 'parent_id': '8134', 'version': '1', 
                         'details': '<p>\n\tDetails of the Test Suite B</p>\n', 'estimated_exec_duration': '3.00', 'updater_id': '2', 'status': '1', 
@@ -199,7 +199,7 @@ SCENARIO_KEYWORDS = {'getTestCasesForTestSuite' : {
                         'steps': [{'step_number': '1', 'actions': '<p>\n\tStep action 1 -b2</p>\n', 'execution_type': '2', 'active': '1', 'id': '8161', 'expected_results': '<p>\n\tStep result 1 - b2</p>\n'}, 
                                   {'step_number': '2', 'actions': '<p>\n\tStep action 2 -b2</p>\n', 'execution_type': '2', 'active': '1', 'id': '8162', 'expected_results': '<p>\n\tStep result 2 - b2</p>\n'}], 
                         'author_id': '2'}, 
-                    {'node_order': '2', 'is_open': '1', 
+                                     {'node_order': '2', 'is_open': '1', 
                      'id': '8169', 'node_type_id': '3', 'layout': '1', 'tc_external_id': '4', 'parent_id': '8134', 'version': '1', 
                      'details': '<p>\n\tDetails of the Test Suite B</p>\n', 'estimated_exec_duration': '3.00', 'updater_id': '2', 'status': '1', 
                      'importance': '3', 'modification_ts': '2014-07-02 21:02:23', 'execution_type': '1', 
@@ -208,8 +208,44 @@ SCENARIO_KEYWORDS = {'getTestCasesForTestSuite' : {
                      'summary': '<p>\n\tThis is the summary of the Test Case B3</p>\n', 
                      'steps': [{'step_number': '1', 'actions': '<p>\n\tStep action 1 -b3</p>\n', 'execution_type': '2', 'active': '1', 'id': '8171', 'expected_results': '<p>\n\tStep result 1 - b3</p>\n'}, 
                                {'step_number': '2', 'actions': '<p>\n\tStep action 2 -b3</p>\n', 'execution_type': '2', 'active': '1', 'id': '8172', 'expected_results': '<p>\n\tStep result 2 - b3</p>\n'}], 
-                     'author_id': '2'}]
-                                            } 
+                     'author_id': '2'}],
+                                                   
+                '4711' :  [{'node_order': '0', 'is_open': '1', 
+                        'keywords': {'1': {'keyword_id': '1', 'notes': 'a key word', 'testcase_id': '8144', 'keyword': 'KeyWord01'}, 
+                                     '3': {'keyword_id': '3', 'notes': 'a third key word', 'testcase_id': '8144', 'keyword': 'KeyWord03'}}, 
+                        'id': '8144', 'node_type_id': '3', 'layout': '1', 'tc_external_id': '2', 'parent_id': '8134', 'version': '1', 
+                        'details': '<p>\n\tDetails of the Test Suite B</p>\n', 'estimated_exec_duration': '3.00', 'updater_id': '2', 'status': '1', 
+                        'importance': '3', 'modification_ts': '2014-06-30 20:45:40', 'execution_type': '1', 
+                        'preconditions': '<p>\n\tthese are the preconditions</p>\n', 'active': '1', 'creation_ts': '2014-06-28 22:06:17', 
+                        'node_table': 'testcases', 'tcversion_id': '8145', 'name': 'TESTCASE_B', 
+                        'summary': '<p>\n\tThis is the summary of the Test Case B</p>\n', 
+                        'steps': [{'step_number': '1', 'actions': 'Step action 1 -b ', 'execution_type': '2', 'active': '1', 'id': '8151', 'expected_results': 'Step result 1 - b'}], 
+                        'author_id': '1'} ],
+
+                'noKeywords' :  [{'node_order': '0', 'is_open': '1', 
+                        'id': '8144', 'node_type_id': '3', 'layout': '1', 'tc_external_id': '2', 'parent_id': '8134', 'version': '1', 
+                        'details': '<p>\n\tDetails of the Test Suite B</p>\n', 'estimated_exec_duration': '3.00', 'updater_id': '2', 'status': '1', 
+                        'importance': '3', 'modification_ts': '2014-06-30 20:45:40', 'execution_type': '1', 
+                        'preconditions': '<p>\n\tthese are the preconditions</p>\n', 'active': '1', 'creation_ts': '2014-06-28 22:06:17', 
+                        'node_table': 'testcases', 'tcversion_id': '8145', 'name': 'TESTCASE_B', 
+                        'summary': '<p>\n\tThis is the summary of the Test Case B</p>\n', 
+                        'steps': [{'step_number': '1', 'actions': 'Step action 1 -b ', 'execution_type': '2', 'active': '1', 'id': '8151', 'expected_results': 'Step result 1 - b'}], 
+                        'author_id': '1'} ]
+
+                                            },
+                     'getTestCase' : {
+                                    '8144' : [{'full_tc_external_id': 'NPROAPI-2', 'id': '8145', 'tc_external_id': '2', 'version': '1', 
+                                                    'testsuite_id': 'deepFalse3', 'testcase_id': '8144', 'name': 'TESTCASE_B'}],
+                                    'NPROAPI-2' : [{'full_tc_external_id': 'NPROAPI-2', 'id': '8145', 'tc_external_id': '2', 'version': '1', 
+                                                    'testsuite_id': 'deepFalse3', 'testcase_id': '8144', 'name': 'TESTCASE_B'}],
+                                    '8159' : [{'full_tc_external_id': 'NPROAPI-3', 'id': '8160', 'tc_external_id': '3', 'version': '1', 
+                                                    'testsuite_id': 'deepFalse3', 'testcase_id': '8159', 'name': 'TESTCASE_B2'}],
+                                    'NPROAPI-3' : [{'full_tc_external_id': 'NPROAPI-3', 'id': '8160', 'tc_external_id': '3', 'version': '1', 
+                                                    'testsuite_id': 'deepFalse3', 'testcase_id': '8159', 'name': 'TESTCASE_B2'}],
+                                    '8169' : [{'full_tc_external_id': 'NPROAPI-4', 'id': '8170', 'tc_external_id': '3', 'version': '1', 
+                                                    'testsuite_id': 'deepFalse3', 'testcase_id': '8169', 'name': 'TESTCASE_B3'}],
+                                    'NPROAPI-4' : [{'full_tc_external_id': 'NPROAPI-4', 'id': '8170', 'tc_external_id': '3', 'version': '1', 
+                                                    'testsuite_id': 'deepFalse3', 'testcase_id': '8169', 'name': 'TESTCASE_B3'}] }
                      }
 
 
@@ -245,7 +281,14 @@ class DummyAPIClient(TestlinkAPIClient):
         elif methodAPI in ['getTestCaseIDByName']:
             response = data[argsAPI['testcasename']]
         elif methodAPI in ['getTestCase']:
-            response = data['%(testcaseid)s-%(version)s' % argsAPI]
+            datakey = argsAPI.get('testcaseid')
+            if datakey:
+                datakey = str(datakey)
+            else:
+                datakey = argsAPI.get('testcaseexternalid', '')
+            if argsAPI.has_key('version'):
+                datakey += '-%(version)s' % argsAPI
+            response = data[datakey]
         elif methodAPI in ['getFullPath']:
             response = data[argsAPI['nodeid']]
         elif methodAPI in ['getTestProjectByName']:
@@ -471,6 +514,57 @@ class TestLinkAPIOfflineTestCase(unittest.TestCase):
     def test_whatArgs_getTestCasesForTestSuite(self):
         argsDescription = self.api.whatArgs('getTestCasesForTestSuite')
         self.assertIn('getkeywords=<getkeywords>', argsDescription)
+        
+    def test_listKeywordsForTC_FullExternalId(self):
+        self.api.loadScenario(SCENARIO_KEYWORDS)
+        response = self.api.listKeywordsForTC('NPROAPI-2')
+        self.assertEqual(['KeyWord01', 'KeyWord03'], response)
+        
+    def test_listKeywordsForTC_InternalId_Int(self):
+        self.api.loadScenario(SCENARIO_KEYWORDS)
+        response = self.api.listKeywordsForTC(8144)
+        self.assertEqual(['KeyWord01', 'KeyWord03'], response)
+        
+    def test_listKeywordsForTC_InternalId_String(self):
+        self.api.loadScenario(SCENARIO_KEYWORDS)
+        response = self.api.listKeywordsForTC('8144')
+        self.assertEqual(['KeyWord01', 'KeyWord03'], response)
+        
+    def test_listKeywordsForTC_One(self):
+        self.api.loadScenario(SCENARIO_KEYWORDS)
+        response = self.api.listKeywordsForTC('NPROAPI-3')
+        self.assertEqual(['KeyWord02'], response)
+
+    def test_listKeywordsForTC_None(self):
+        self.api.loadScenario(SCENARIO_KEYWORDS)
+        response = self.api.listKeywordsForTC('NPROAPI-4')
+        self.assertEqual([], response)
+        
+    def test_listKeywordsForTS_NoneTC(self):
+        self.api.loadScenario(SCENARIO_KEYWORDS)
+        response = self.api.listKeywordsForTS('noTestCase')
+        self.assertEqual({}, response)
+
+    def test_listKeywordsForTS_NoneKW(self):
+        self.api.loadScenario(SCENARIO_KEYWORDS)
+        response = self.api.listKeywordsForTS('noKeywords')
+        self.assertEqual({'8144' : []}, response)
+        
+    def test_listKeywordsForTS_Id_Int(self):
+        self.api.loadScenario(SCENARIO_KEYWORDS)
+        response = self.api.listKeywordsForTS(4711)
+        self.assertEqual({'8144' : ['KeyWord01', 'KeyWord03']}, response)
+        
+    def test_listKeywordsForTS_Id_String(self):
+        self.api.loadScenario(SCENARIO_KEYWORDS)
+        response = self.api.listKeywordsForTS('4711')
+        self.assertEqual({'8144' : ['KeyWord01', 'KeyWord03']}, response)
+       
+    def test_listKeywordsForTS_Multi(self):
+        self.api.loadScenario(SCENARIO_KEYWORDS)
+        response = self.api.listKeywordsForTS('deepFalse3')
+        self.assertEqual({'8144' : ['KeyWord01', 'KeyWord03'],
+                          '8159' : ['KeyWord02'], '8169' : []}, response)